backend/requirements.txt
Django==5.0.1
django-jazzmin>=2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
Pillow>=10.4.0
python-dotenv==1.0.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
whitenoise==6.6.0
gunicorn==21.2.0
psycopg[binary]>=3.1.0
